The dynamic backend feature causes crashes
- Added "R" to the short options
- Added "backend-refresh" to config file parsing
- Made "backend-refresh" default to 0.
- Started to use suckaddr for the convenient compare, copy, etc functions
- Storing payload as sockaddr_storage as the size of suckaddr in not available
  compile time
- Man pages update

+++ src/tests/test25-dynamic-backend-address.sh
@@ -11,6 +11,7 @@ mk_cfg <<EOF
 pem-file = "${CERTSDIR}/default.example.com"
 frontend = "[$LISTENADDR]:$LISTENPORT"
 backend = "[hitch_test]:80"
+backend-refresh = 1
 EOF
 
 hitch $HITCH_ARGS --config=$CONFFILE
@@ -29,7 +30,11 @@ runcurl $LISTENADDR $LISTENPORT
 
 curl --max-time 5  --insecure https://$LISTENADDR:$LISTENPORT/ >$DUMPFILE 2>&1
 test $? -eq 0 || die "New backend address should be available."
-grep -q -c "We moved the Varnish Project to a new server" $DUMPFILE
+
+grep "<title>Hitch TLS proxy" $DUMPFILE
+test $? -eq 1 || die "Old backend address should NOT be available."
+
+grep "We moved the Varnish Project to a new server" $DUMPFILE
 test $? -eq 0 || die "New backend served unexpected site"
 
 rm $TMP_HOSTS_FILE
